refactor: remove excessive mw
